+++ Communication.h
@@ -308,10 +308,10 @@ public:
     static void printFLN(FSTRINGPARAM(text), uint16_t value);
     static void printFLN(FSTRINGPARAM(text), int32_t value);
     static void printFLN(FSTRINGPARAM(text), uint32_t value);
-    static void printFLN(FSTRINGPARAM(text), const char *msg);
+    static void printFLN(FSTRINGPARAM(text), const char* msg);
     static void printFLN(FSTRINGPARAM(text), float value,uint8_t digits=2);
-    static void printArrayFLN(FSTRINGPARAM(text), float *arr, uint8_t n=4, uint8_t digits=2);
-    static void printArrayFLN(FSTRINGPARAM(text), long *arr, uint8_t n=4);
+    static void printArrayFLN(FSTRINGPARAM(text), float* arr, uint8_t n=4, uint8_t digits=2);
+    static void printArrayFLN(FSTRINGPARAM(text), int32_t* arr, uint8_t n=4);
     static void print(int32_t value);
     static inline void print(uint32_t value) { printNumber(value); }
     static inline void print(int16_t value) { printNumber((int32_t)value); }
@@ -330,7 +330,7 @@ private:
 #ifdef DEBUG
 #define SHOW(x)                 {Com::printF(PSTR(" " #x "=")); Com::print(x); Com::println();}
 #define SHOWS(x)                {Com::printF(PSTR(" " #x "=")); Com::print(x); Com::print(" steps  "); Com::print(x/80); Com::printFLN(PSTR(" mm"));}
-#define SHOWM(x)                {Com::printF(PSTR(" " #x "=")); Com::print((long)x*80); Com::print(" steps  "); Com::print(x); Com::printFLN(PSTR(" mm"));}
+#define SHOWM(x)                {Com::printF(PSTR(" " #x "=")); Com::print((int32_t)x*80); Com::print(" steps  "); Com::print(x); Com::printFLN(PSTR(" mm"));}
 #define SHOT(x)                 Com::printF(PSTR(x " "))
 #define SHOWA(t,a,n)            {SHOT(t); for (int i=0;i<n;i++) SHOWS(a[i]);}
 #define SHOWAM(t,a,n)           {SHOT(t); for (int i=0;i<n;i++) SHOWM(a[i]);}
